moved statusAnnouncer
