+++ packages/model-viewer/src/three-components/StatusAnnouncer.ts
@@ -14,8 +14,8 @@
  */
 
 import {EventDispatcher} from 'three';
-import ModelViewerElementBase from '../../model-viewer-base.js';
-import {debounce, getFirstMapKey} from '../../utilities.js';
+import ModelViewerElementBase from '../model-viewer-base.js';
+import {debounce, getFirstMapKey} from '../utilities.js';
 
 export const INITIAL_STATUS_ANNOUNCEMENT =
     'This page includes one or more 3D models that are loading';
